C_psittaci_folder/scripts/requirements.txt
biopython==1.81
contourpy==1.0.7
cycler==0.11.0
ete3==3.1.2
fonttools==4.39.3
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
seaborn==0.12.2
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
zipp==3.15.0
